LCY : 12-08 17:06 ProductController 완료 PurchaseController 진행중
ProductMapper.xml add 수정

+++ src/main/java/com/cornstory/controller/chat/ChatController.java
@@ -108,15 +108,17 @@ public class ChatController {
             try {
                 // 기존 파일 삭제
                 String deleteImg = chatService.getChatSpace(chatSpace.getChatSpaceNo()).getcSpaceImage();
-                String deleteDir = uploadDir + File.separator + deleteImg;
-                File fileToDelete = new File(deleteDir);
+                if (!deleteImg.equals("chat.jpg")) {
+                    String deleteDir = uploadDir + File.separator + deleteImg;
+                    File fileToDelete = new File(deleteDir);
 
-                // 파일을 삭제합니다.
-                if (fileToDelete.exists()) {
-                    if (fileToDelete.delete()) {
-                        System.out.println("파일이 성공적으로 삭제되었습니다.");
-                    } else {
-                        System.out.println("파일을 삭제하는 데 문제가 발생했습니다.");
+                    // 파일을 삭제합니다.
+                    if (fileToDelete.exists()) {
+                        if (fileToDelete.delete()) {
+                            System.out.println("파일이 성공적으로 삭제되었습니다.");
+                        } else {
+                            System.out.println("파일을 삭제하는 데 문제가 발생했습니다.");
+                        }
                     }
                 }
