+++ src/main/java/com/cornstory/controller/product/ProductController.java
@@ -46,9 +46,16 @@ public class ProductController {
     }
 
     @GetMapping("addProduct")
-    public String addProduct(Model model, @RequestParam("prodCategory") int prodCategory) throws Exception {
+    public String addProduct(Model model, @RequestParam("prodCategory") int prodCategory,
+                             HttpSession session) throws Exception {
         System.out.println("/product/addProduct : GET :: prodCategory :: "+prodCategory);
         model.addAttribute("prodCategory", prodCategory);
+
+        if (prodCategory == 2) {
+            List<Work> workList = productService.listCompleteWork(((User) session.getAttribute("user")).getUserId());
+            System.out.println("/product/addProduct : " + workList);
+            model.addAttribute("workList", workList);
+        }
 
         return "product/addProduct";
     }
@@ -81,14 +88,12 @@ public class ProductController {
             product.setProdName(work.getWorkName()+" "+episode.getEpisodeOrder()+" 회차");
         } else if (product.getProdCategory() == 2) {
             product.setProdName(work.getWorkName()+" 저작권");
-            List<Work> workList = productService.listCompleteWork((String) session.getAttribute("userId"));
-            model.addAttribute("workList", workList);
         }
         System.out.println("/product/addProduct : " + product);
 
 
         // https://action713.tistory.com/entry/%EC%8A%A4%ED%94%84%EB%A7%81-%ED%8C%8C%EC%9D%BC-%EA%B2%BD%EB%A1%9C
-        String uploadDir = request.getServletContext().getRealPath("")+"\\..\\resources\\static\\file\\chat\\";
+        String uploadDir = request.getServletContext().getRealPath("")+"\\..\\resources\\static\\file\\product\\";
         // request.getServletContext().getRealPath(""): webapp 상대 경로
 
         if (product.getProdCategory() == 0) {
@@ -111,10 +116,14 @@ public class ProductController {
                 }
             } else {
                 // 업로드된 파일이 없는 경우 처리
-                product.setProdImage("popcorn.png");
+                product.setProdImage("popcorn.jpg");
             }
         }
-        return "redirect:/product/listStore";
+
+        //Business Logic
+        productService.addProduct(product);
+
+        return "redirect:/product/listProduct";
     }
 
     @GetMapping("updateProduct")
@@ -138,22 +147,24 @@ public class ProductController {
 
 
         // https://action713.tistory.com/entry/%EC%8A%A4%ED%94%84%EB%A7%81-%ED%8C%8C%EC%9D%BC-%EA%B2%BD%EB%A1%9C
-        String uploadDir = request.getServletContext().getRealPath("")+"\\..\\resources\\static\\file\\chat\\";
+        String uploadDir = request.getServletContext().getRealPath("")+"\\..\\resources\\static\\file\\product\\";
         // request.getServletContext().getRealPath(""): webapp 상대 경로
         if (product.getProdCategory() == 0) {
             if (!file.isEmpty()) {
                 try {
                     // 기존 파일 삭제
                     String deleteImg = productService.getProduct(product.getProdNo()).getProdImage();
-                    String deleteDir = uploadDir + File.separator + deleteImg;
-                    File fileToDelete = new File(deleteDir);
+                    if (!deleteImg.equals("popcorn.jpg")) {
+                        String deleteDir = uploadDir + File.separator + deleteImg;
+                        File fileToDelete = new File(deleteDir);
 
-                    // 파일을 삭제합니다.
-                    if (fileToDelete.exists()) {
-                        if (fileToDelete.delete()) {
-                            System.out.println("파일이 성공적으로 삭제되었습니다.");
-                        } else {
-                            System.out.println("파일을 삭제하는 데 문제가 발생했습니다.");
+                        // 파일을 삭제합니다.
+                        if (fileToDelete.exists()) {
+                            if (fileToDelete.delete()) {
+                                System.out.println("파일이 성공적으로 삭제되었습니다.");
+                            } else {
+                                System.out.println("파일을 삭제하는 데 문제가 발생했습니다.");
+                            }
                         }
                     }
 
@@ -178,7 +189,10 @@ public class ProductController {
             }
         }
 
-        return "redirect:/product/listStore";
+        //Business Logic
+        productService.updateProduct(product);
+
+        return "redirect:/product/listProduct";
     }
 
     @RequestMapping("listProduct")
